+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- IndieAuth: added `context.Context` as argument to all functions that perform HTTP requests.
+
 ### Deprecated
 
 ### Removed